requirements.txt
# You can optionally pin to a version of the Runway Model SDK like: "runway-python==0.1.0"
runway-python
pillow>=6.2.0
opencv-contrib-python